Fix Makefile to use the PGXS's CFLAGS properly.
In the previous coding, we didn't actually inherit any flags from PGXS.
One of the significant flag was -g, which sometimes made it difficult
to debug the code.

+++ plv8_param.cc
@@ -50,7 +50,7 @@ plv8_variable_paramref_hook(ParseState *pstate, ParamRef *pref)
 	Param	   *param;
 
 	/* Check parameter number is in range */
-	if (paramno <= 0 || paramno > INT_MAX / sizeof(Oid))
+	if (paramno <= 0 || paramno > (int) (INT_MAX / sizeof(Oid)))
 		ereport(ERROR,
 				(errcode(ERRCODE_UNDEFINED_PARAMETER),
 				 errmsg("there is no parameter $%d", paramno),
